+++ sampleproject/examples/urls.py
@@ -34,7 +34,8 @@ def get_example_urls():
             view_class = None
             for attr_name in dir(module):
                 attr = getattr(module, attr_name)
-                if issubclass(attr, Component) and attr_name != "Component" and attr_name.endswith("Page"):
+                is_component_sublass = isinstance(attr, type) and issubclass(attr, Component)
+                if is_component_sublass and attr_name != "Component" and attr_name.endswith("Page"):
                     view_class = attr
                     break
 
@@ -50,6 +51,7 @@ def get_example_urls():
 
         except Exception as e:  # noqa: BLE001
             print(f"Failed to register URL for {example_name}: {e}")
+            raise
 
     return urlpatterns